Create LC_API_call
Queries LC API for rows with AUTH Originating System ID. Note that this function only works on LCNames; could be improved by excluding lines with non-standard LCCNs (e.g. ones that don't start with n) and lines where vocabulary is LCSH. 
Major enhancement would be to query API based on heading string, but this would require more parsing of results to note uncertainty of matching; may not be possible or desirable.